backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
email-validator
apscheduler
twilio
